requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai>=1.0.0
chromadb==0.4.22
openpyxl==3.1.2
pandas==2.1.4
python-dotenv==1.0.0
pydantic==2.5.3
pyyaml==6.0.1
numpy>=1.24.0
